Restructure Layout
This commit restructures the layout so that we can easily add some
global styles to the app and include the Google Fonts required in the 
header.

Using `next/Head` this is fairly easy, although for performance reasons
it's probably not optimal.

To add some styles to the Header (which moved to be part of layout) CSS
modules has been introduced. This would preferably been a separate
commit but splitting every single logical thing could take hours so
progress is now my aim!