Introduce FormState for Name field
The idea is that the form will essential be controlled by the basic
state machine `INITIAL_STATE` which will decide page content, handle
validation and errors, pass all the relevant information to the form
inputs and also control the journey through the form.

With this structure we can simply add to the `INITIAL_STATE` object and
further form fields on their own pages will 'just work'.

This will make testing easier as we can basically decide on the contract
between the from object and what should be displayed in the specs very
easily (easier than getting jest working with NextJS).

The handleChange function should work for any type of input, as the
`event` object is fairly standard for getting values etc from any form
field.

Initially this will only work with the `InputWithLabel` component but it
should be possible to enable a Radio button group later.