Add validation for Name field
This adds validation for the Name field, using a separate function we
can abstract away if needed.

We return a string for a simple API that we can use for all future
fields, because we can do on the fly validation we should only ever need
to display one error message so a single string should be fine.

And by attaching it to the state machine, we can define a different
function for each field and do specific validations however we like.