Add Email Step
This uses the form initial state machine to add the email step.

By simply adding to the 'model' we have created in the INITIAL_STATE
object we can just build on the existing set up and it all will work.

No validation exists for the email field but that will follow in a
subsequent commit.