Add Disabled prop to Button component
I noticed that the page broke when next Step wasn't defined. This was
kind of expected but not completely broken. This adds a prop to allow
the button to be disabled if it is unusable and associated styles to
indicate that the button can not be clicked. The colour contrast is not
good for accessibility but as this is just a code test, I'm not spending
time on designing a disable button state.