Allow users to step back and forwards when page available
This adds a `prevStep` item to the Form State to control which page the
user goes back to. This could be done by managing the state of which
page was previously shown but this API is simpler, if more prone to
human error.

The buttons can now take `nextStep` or `prevStep` to be disabled if there's
no page to navigate to, and we can now enable the back button with a simple
currentStep change.