Add Service Form step
This adds the Radio button question for what service are you interested
in.  The form state machine set up makes it easy to add the required
data for the radio buttons, albeit with a slightly different API to
an input field.

The more complex bit is switching to a different component for different
types of question. This is done by attaching a component to each form
step and assigning that to a `FormComponent` constant in the main
component. This could have been worked out through the shape of the
currentStep, but this is a simpler solution for this exercise.